refactored permission request workflow
