+++ app/src/main/java/com/test/dvt/presentation/navigation/StartNavigationGraph.kt
@@ -1,199 +1,21 @@
 package com.test.dvt.presentation.navigation
 
-import android.Manifest
 import android.annotation.SuppressLint
-import androidx.compose.animation.animateContentSize
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxScope
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.google.accompanist.permissions.ExperimentalPermissionsApi
-import com.google.accompanist.permissions.MultiplePermissionsState
-import com.google.accompanist.permissions.isGranted
-import com.google.accompanist.permissions.rememberMultiplePermissionsState
 import com.test.dvt.presentation.screens.SplashScreen
 
 @SuppressLint("MissingPermission")
 @Composable
 fun StartNavigationGraph(navController: NavHostController) {
-    val permissions = listOf(
-        Manifest.permission.ACCESS_COARSE_LOCATION,
-        Manifest.permission.ACCESS_FINE_LOCATION,
-    )
     NavHost(
         navController = navController,
         startDestination = BottomNavRoutes.Splash.route
     ) {
         composable(route = BottomNavRoutes.Splash.route) {
-            PermissionBox(
-                permissions = permissions,
-                requiredPermissions = listOf(permissions.first()),
-                onGranted = {
-                    SplashScreen(usePreciseLocation = it.contains(Manifest.permission.ACCESS_FINE_LOCATION))
-                },
-            )
+            SplashScreen()
         }
-    }
-}
-
-@OptIn(ExperimentalPermissionsApi::class)
-@Composable
-fun PermissionBox(
-    modifier: Modifier = Modifier,
-    permissions: List<String>,
-    requiredPermissions: List<String> = permissions,
-    description: String? = null,
-    contentAlignment: Alignment = Alignment.TopStart,
-    onGranted: @Composable BoxScope.(List<String>) -> Unit,
-) {
-    var errorText by remember {
-        mutableStateOf("")
-    }
-
-    val permissionState = rememberMultiplePermissionsState(permissions = permissions) { map ->
-        val rejectedPermissions = map.filterValues { !it }.keys
-        errorText = if (rejectedPermissions.none { it in requiredPermissions }) {
-            ""
-        } else {
-            "${rejectedPermissions.joinToString()} required for the sample"
-        }
-    }
-    val allRequiredPermissionsGranted =
-        permissionState.revokedPermissions.none { it.permission in requiredPermissions }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .then(modifier),
-        contentAlignment = if (allRequiredPermissionsGranted) {
-            contentAlignment
-        } else {
-            Alignment.Center
-        },
-    ) {
-        if (allRequiredPermissionsGranted) {
-            onGranted(
-                permissionState.permissions
-                    .filter { it.status.isGranted }
-                    .map { it.permission },
-            )
-        } else {
-            PermissionScreen(
-                permissionState,
-                description,
-                errorText,
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalPermissionsApi::class)
-@Composable
-private fun PermissionScreen(
-    state: MultiplePermissionsState,
-    description: String?,
-    errorText: String,
-) {
-    var showRationale by remember(state) {
-        mutableStateOf(false)
-    }
-
-    val permissions = remember(state.revokedPermissions) {
-        state.revokedPermissions.joinToString("\n") {
-            " - " + it.permission.removePrefix("android.permission.")
-        }
-    }
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .animateContentSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Text(
-            text = "Sample requires permission/s:",
-            style = MaterialTheme.typography.titleLarge,
-            modifier = Modifier.padding(16.dp),
-        )
-        Text(
-            text = permissions,
-            style = MaterialTheme.typography.bodyMedium,
-            modifier = Modifier.padding(16.dp),
-        )
-        if (description != null) {
-            Text(
-                text = description,
-                style = MaterialTheme.typography.bodyMedium,
-                modifier = Modifier.padding(16.dp),
-            )
-        }
-        Button(
-            onClick = {
-                if (state.shouldShowRationale) {
-                    showRationale = true
-                } else {
-                    state.launchMultiplePermissionRequest()
-                }
-            },
-        ) {
-            Text(text = "Grant permissions")
-        }
-        if (errorText.isNotBlank()) {
-            Text(
-                text = errorText,
-                style = MaterialTheme.typography.labelSmall,
-                modifier = Modifier.padding(16.dp),
-            )
-        }
-    }
-    if (showRationale) {
-        AlertDialog(
-            onDismissRequest = {
-                showRationale = false
-            },
-            title = {
-                Text(text = "Permissions required by the sample")
-            },
-            text = {
-                Text(text = "The sample requires the following permissions to work:\n $permissions")
-            },
-            confirmButton = {
-                TextButton(
-                    onClick = {
-                        showRationale = false
-                        state.launchMultiplePermissionRequest()
-                    },
-                ) {
-                    Text("Continue")
-                }
-            },
-            dismissButton = {
-                TextButton(
-                    onClick = {
-                        showRationale = false
-                    },
-                ) {
-                    Text("Dismiss")
-                }
-            },
-        )
     }
 }
